feat: 대댓글 관리 기능 (#268)
* feat: 댓글 수정기능 추가

* feat: Cocomment 등록 기능 추가 및 Comment Update 기능 관련 테스트 추가

* test: 대댓글 생성 기능 서비스 테스트 추가

* test: 대댓글 작성기능 인수테스트 추가

* test: 대댓글, 댓글 순차 삭제의 경우 인수테스트 추가

* feat: CommentState 추가

* refactor: 댓글/대댓글 삭제 기능 개선

* feat: Flyway 반영

- Local & Test도 Flyway 가능하게끔 변경
- User Email 삭제되더라도 무조건 Unique하게 DB상 설정하여 예외 발생시 처리하는 로직 추가
- flyway DB Table 추가
- application.properties 정비
- SpringBootTest RandomPort 이용하도록 변경

* refactor: application.properties 속성 수정

- enable_lazy_load_no_trans 사용하지 않는 것이 좋다고 판단하여 제거(안티패턴)
- spring.datasource.driver-class-name, username, password 기본 값과 동일하여 제거

* refactor: Flyway SQL명 이슈번호 추가 & 사용하지 않는 설정 제거

* refactor: 오타 수정


* refactor: 오타 수정


* refactor: 파일명 변경

+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/service/CommentService.java
@@ -1,6 +1,10 @@
 package wooteco.team.ittabi.legenoaroundhere.service;
 
+import static wooteco.team.ittabi.legenoaroundhere.domain.State.DELETED;
+
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -12,7 +16,9 @@ import wooteco.team.ittabi.legenoaroundhere.domain.user.User;
 import wooteco.team.ittabi.legenoaroundhere.dto.CommentAssembler;
 import wooteco.team.ittabi.legenoaroundhere.dto.CommentRequest;
 import wooteco.team.ittabi.legenoaroundhere.dto.CommentResponse;
+import wooteco.team.ittabi.legenoaroundhere.dto.CommentResponseAssembler;
 import wooteco.team.ittabi.legenoaroundhere.exception.NotAuthorizedException;
+import wooteco.team.ittabi.legenoaroundhere.exception.NotAvailableException;
 import wooteco.team.ittabi.legenoaroundhere.exception.NotExistsException;
 import wooteco.team.ittabi.legenoaroundhere.repository.CommentRepository;
 import wooteco.team.ittabi.legenoaroundhere.repository.PostRepository;
@@ -36,7 +42,7 @@ public class CommentService {
         comment.setPost(post);
 
         Comment savedComment = commentRepository.save(comment);
-        return CommentResponse.of(user, savedComment);
+        return CommentResponseAssembler.of(user, savedComment);
     }
 
     @Transactional(readOnly = true)
@@ -44,22 +50,40 @@ public class CommentService {
         User user = (User) authenticationFacade.getPrincipal();
 
         Comment comment = findCommentBy(commentId);
-        return CommentResponse.of(user, comment);
+        return CommentResponseAssembler.of(user, comment);
     }
 
     @Transactional(readOnly = true)
     public List<CommentResponse> findAllCommentBy(Long postId) {
         User user = (User) authenticationFacade.getPrincipal();
 
-        List<Comment> comments = commentRepository.findAllByPostId(postId);
-        return CommentResponse.listOf(user, comments);
+        List<Comment> comments = commentRepository.findAllByPostId(postId)
+            .stream()
+            .filter(Comment::hasSuperComment)
+            .collect(Collectors.toList());
+
+        return CommentResponseAssembler.listOf(user, comments);
     }
 
     @Transactional
     public void deleteComment(Long commentId) {
-        Comment comment = findCommentBy(commentId);
+        Comment comment = findAvailableCommentBy(commentId);
         validateIsCreator(comment);
+        if (comment.hasCocomments()) {
+            comment.setState(DELETED);
+            return;
+        }
+        if (isSuperCommentNotAvailableAndOnlyCocomment(comment)) {
+            commentRepository.delete(comment.getSuperComment());
+        }
         commentRepository.delete(comment);
+    }
+
+    private boolean isSuperCommentNotAvailableAndOnlyCocomment(Comment comment) {
+        Comment superComment = comment.getSuperComment();
+        return Objects.nonNull(superComment)
+            && superComment.isNotAvailable()
+            && superComment.hasOnlyCocomment();
     }
 
     private Comment findCommentBy(Long commentId) {
@@ -80,7 +104,38 @@ public class CommentService {
     public void pressZzang(Long commentId) {
         User user = (User) authenticationFacade.getPrincipal();
 
-        Comment comment = findCommentBy(commentId);
+        Comment comment = findAvailableCommentBy(commentId);
         comment.pressZzang(user);
+    }
+
+    @Transactional
+    public CommentResponse updateComment(Long commentId, CommentRequest commentRequest) {
+        User user = (User) authenticationFacade.getPrincipal();
+        Comment comment = findAvailableCommentBy(commentId);
+
+        comment.setWriting(commentRequest.getWriting());
+
+        return CommentResponseAssembler.of(user, comment);
+    }
+
+    @Transactional
+    public CommentResponse createCocomment(Long commentId, CommentRequest commentRequest) {
+        User user = (User) authenticationFacade.getPrincipal();
+
+        Comment comment = findAvailableCommentBy(commentId);
+        Comment cocomment = CommentAssembler.assemble(user, commentRequest);
+        cocomment.setSuperComment(comment);
+
+        Comment savedCocomment = commentRepository.save(cocomment);
+        return CommentResponseAssembler.of(user, savedCocomment);
+    }
+
+    private Comment findAvailableCommentBy(Long commentId) {
+        Comment comment = findCommentBy(commentId);
+        if (comment.isNotAvailable()) {
+            throw new NotAvailableException(
+                "Comment ID : " + comment.getId() + " 에 해당하는 Comment가 유효하지 않습니다!");
+        }
+        return comment;
     }
 }
